+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+###
+- added TryGetValue to NavigationParameters
+
+### Changed
+- change GoBack to check if there are pages on modal stack and pop modal page if there is one
 
 ## [0.2.0] - 2023-11-14